Modified vehicle groups on change ownership added guard condition

+++ dotnet/corecomponents/net.atos.daf.ct2.vehicle/repository/VehicleRepository.cs
@@ -575,37 +575,40 @@ namespace net.atos.daf.ct2.vehicle.repository
                 parameter.Add("@is_ota", vehicle.Is_Ota);
                 int vehicleID = await _dataAccess.ExecuteScalarAsync<int>(queryStatement, parameter);
 
-                //Update existing vehicle groups if change ownership happens
-                //Delete single groups and access relationships created in old owner org
-                parameter = new DynamicParameters();
-                parameter.Add("@Id", vehicle.ID);
-                parameter.Add("@Old_OrganizationId", vehicleDetails.OrganizationId);
+                if (vehicle.Organization_Id != vehicleDetails.OrganizationId)
+                {
+                    //Update existing vehicle groups if change ownership happens
+                    //Delete single groups and access relationships created in old owner org
+                    parameter = new DynamicParameters();
+                    parameter.Add("@Id", vehicle.ID);
+                    parameter.Add("@Old_OrganizationId", vehicleDetails.OrganizationId);
 
-                //Check if access relationship exists on the vehicle and old owner org with single vehicle group type
-
-                var vehicleGroupIds = await _dataAccess
-                                    .QueryAsync<int>(@"select vehicle_group_id from master.accessrelationship where vehicle_group_id in 
+                    //Check if access relationship exists on the vehicle and old owner org with single vehicle group type
+                    var vehicleGroupIds = await _dataAccess
+                                        .QueryAsync<int>(@"select vehicle_group_id from master.accessrelationship where vehicle_group_id in 
                                                    (select id from master.group where ref_id = @Id 
                                                     and organization_id = @Old_OrganizationId
                                                     and group_type = 'S' and object_type = 'V')", parameter);
 
-                if (vehicleGroupIds.Count() > 0)
-                {
-                    var parameters = new DynamicParameters();
-                    parameters.Add("@VehicleGroupIds", vehicleGroupIds.ToArray());
-                    //Delete access relationships with single vehicle group
-                    await _dataAccess.ExecuteAsync(@"delete from master.master.accessrelationship where vehicle_group_id = ANY(@VehicleGroupIds)", parameter);
-                }
+                    if (vehicleGroupIds.Count() > 0)
+                    {
+                        var parameters = new DynamicParameters();
+                        parameters.Add("@VehicleGroupIds", vehicleGroupIds.ToArray());
+                        //Delete access relationships with single vehicle group
+                        await _dataAccess.ExecuteAsync(@"delete from master.master.accessrelationship where vehicle_group_id = ANY(@VehicleGroupIds)", parameter);
+                    }
 
-                //Delete groups with single type
-                await _dataAccess.ExecuteAsync(@"delete from master.group where ref_id = @Id and organization_id = @Old_OrganizationId 
+                    //Delete groups with single type
+                    await _dataAccess.ExecuteAsync(@"delete from master.group where ref_id = @Id and organization_id = @Old_OrganizationId 
                                                  and group_type = 'S' and object_type = 'V'", parameter);
 
-                //Delete vehicle entries from GroupRef for old owner org
-                await _dataAccess.ExecuteAsync(@"delete from master.groupref gref
+                    //Delete vehicle entries from GroupRef for old owner org
+                    await _dataAccess.ExecuteAsync(@"delete from master.groupref gref
                                                 where gref.ref_id = @Id and gref.group_id in 
                                                     (select id from master.group grp where grp.organization_id = @Old_OrganizationId 
                                                     and grp.group_type = 'G' and grp.object_type = 'V')", parameter);
+
+                }
 
                 transaction.Commit();
                 return vehicle;
